backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
pdfplumber==0.10.3
python-docx==1.1.0
thefuzz==0.22.1
nltk==3.8.1
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0